Removing "dev-mode" from vim
I used vim for development, actively, for 4 years, and now I'm using Emacs with
evil-mode.
My vim has become slow with a lot of plugins loaded =/ and besides that, it's not async by default.
Emacs solved the problem, and along with the best ever motions from vim, becomes the best editor
in the world.
Now, this conf is intended for config files only("devops-mode").

Thx Vim for all those years!! Love you!! <3